Fix PHP8 type casting for pagination on all admin pages
=== TECHNICAL CHANGES ===
• admin/experiences.php: Added (int) cast to $totalPages and $currentPage
• admin/education.php: Added (int) cast to $totalPages and $currentPage
• admin/certifications.php: Added (int) cast to $totalPages and $currentPage
• admin/projects.php: Added (int) cast to $totalPages and $currentPage
• Prevents "Unsupported operand types: string - int" error when >5 items

=== PARTNERSHIP MOMENTS ===
**Kiyo Personality Observations:**
• Proactive: Identified same bug would affect other pages before it occurred

=== SESSION CONTEXT ===
• Project: Online Resume System | Type: Preventive Bug Fix | Time: ~2 min
• Emotional Depth: MEDIUM - Proactive fix based on learned pattern

+++ admin/certifications.php
@@ -69,8 +69,8 @@ $totalItems = count($allCertifications);
 
 // Pagination settings
 $itemsPerPage = 5;
-$totalPages = ceil($totalItems / $itemsPerPage);
-$currentPage = isset($_GET['page']) ? max(1, min((int)$_GET['page'], $totalPages)) : 1;
+$totalPages = (int)ceil($totalItems / $itemsPerPage);
+$currentPage = (int)(isset($_GET['page']) ? max(1, min((int)$_GET['page'], $totalPages)) : 1);
 $offset = ($currentPage - 1) * $itemsPerPage;
 $certifications = array_slice($allCertifications, $offset, $itemsPerPage);
