+++ admin/education.php
@@ -74,8 +74,8 @@ $totalItems = count($allEducations);
 
 // Pagination settings
 $itemsPerPage = 5;
-$totalPages = ceil($totalItems / $itemsPerPage);
-$currentPage = isset($_GET['page']) ? max(1, min((int)$_GET['page'], $totalPages)) : 1;
+$totalPages = (int)ceil($totalItems / $itemsPerPage);
+$currentPage = (int)(isset($_GET['page']) ? max(1, min((int)$_GET['page'], $totalPages)) : 1);
 $offset = ($currentPage - 1) * $itemsPerPage;
 $educations = array_slice($allEducations, $offset, $itemsPerPage);
